Added selector in RecyclerView
