+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.3"
     defaultConfig {
-        applicationId "app.com.warattil"
+        applicationId "com.crownstack.surah"
         minSdkVersion 16
         targetSdkVersion 25
         versionCode 1
