+++ app/src/main/java/app/com/warattil/activities/SplashActivity.java
@@ -20,7 +20,6 @@ public class SplashActivity extends AppCompatActivity implements Constants {
     @BindView(R.id.image_view_splash) ImageView mImageViewSplash;
 
     private static final long FIRST_SPLASH_TIME = 2000;
-    private static final long SECOND_SPLASH_TIME = 1000;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,32 +31,13 @@ public class SplashActivity extends AppCompatActivity implements Constants {
             @Override
             public void run() {
 
-                Drawable drawable = ContextCompat.getDrawable(getApplicationContext(), R.drawable.splash_second);
-                mImageViewSplash.setBackground(drawable); // this method require min sdk 16
                 nextActivity();
-
             }
         }, FIRST_SPLASH_TIME);
     }
 
     private void nextActivity() {
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                retrievePreference();
-            }
-        }, SECOND_SPLASH_TIME);
-    }
-
-    private void retrievePreference() {
-        String storedLanguage = AppPreference.getAppPreference(SplashActivity.this).getString(PREF_LANGUAGE);
-
-        if (storedLanguage.equals(PREF_LANGUAGE_ENGLISH) || storedLanguage.equals(PREF_LANGUAGE_ARABIC) ) {
-            startActivity(new Intent(getApplicationContext(), SongListActivity.class));
-            finish();
-        } else {
-            startActivity(new Intent(getApplicationContext(), SettingActivity.class));
-            finish();
-        }
+       startActivity(new Intent(this, SplashSecondActivity.class));
+        finish();
     }
 }
