+++ app/src/main/res/drawable/selector_row.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- Color when the row is selected -->
+    <item android:state_pressed="true">
+        <shape>
+            <solid android:color="@color/colorCoralTree"/>
+        </shape>
+    </item>
+    <!-- Standard background color -->
+    <item android:state_pressed="false">
+        <shape>
+            <solid android:color="@android:color/transparent"/>
+        </shape>
+    </item>
+</selector>
