+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Warattil</string>
+    <string name="app_name">Surah</string>
     <string name="language">Language</string>
     <string name="english">English</string>
     <string name="arabic">Arabic</string>
@@ -22,4 +22,5 @@
     <string name="select_language_reciter">Please select a language and a reciter</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="check_connection">Please check your internet connection.</string>
+    <string name="splash_text">Surah \n  سورة</string>
 </resources>
